fix(cli): Fix `fetchLibraryVersions` function import
