+++ packages/cli/utils.js
@@ -1,3 +1,5 @@
+const algoliasearch = require('algoliasearch');
+
 function camelCase(string) {
   return string.replace(/-([a-z])/g, str => str[1].toUpperCase());
 }
@@ -38,8 +40,24 @@ function isQuestionAsked({ question, args }) {
   return true;
 }
 
+const algoliaConfig = {
+  appId: 'OFCNCOG2CU',
+  apiKey: 'f54e21fa3a2a0160595bb058179bfb1e',
+  indexName: 'npm-search',
+};
+
+const client = algoliasearch(algoliaConfig.appId, algoliaConfig.apiKey);
+const index = client.initIndex(algoliaConfig.indexName);
+
+async function fetchLibraryVersions(libraryName) {
+  const library = await index.getObject(libraryName);
+
+  return Object.keys(library.versions).reverse();
+}
+
 module.exports = {
   camelCase,
   getOptionsFromArguments,
   isQuestionAsked,
+  fetchLibraryVersions,
 };
