+++ packages/shared/utils.js
@@ -2,7 +2,6 @@ const fs = require('fs');
 const { execSync } = require('child_process');
 const chalk = require('chalk');
 const validateProjectName = require('validate-npm-package-name');
-const algoliasearch = require('algoliasearch');
 
 function checkAppName(name) {
   const validationResult = validateProjectName(name);
@@ -62,26 +61,10 @@ function isYarnAvailable() {
   }
 }
 
-const algoliaConfig = {
-  appId: 'OFCNCOG2CU',
-  apiKey: 'f54e21fa3a2a0160595bb058179bfb1e',
-  indexName: 'npm-search',
-};
-
-const client = algoliasearch(algoliaConfig.appId, algoliaConfig.apiKey);
-const index = client.initIndex(algoliaConfig.indexName);
-
-async function fetchLibraryVersions(libraryName) {
-  const library = await index.getObject(libraryName);
-
-  return Object.keys(library.versions).reverse();
-}
-
 module.exports = {
   checkAppName,
   checkAppPath,
   isYarnAvailable,
   getTemplateName,
   getLibraryName,
-  fetchLibraryVersions,
 };
